feat: solve nodules task
